+++ src/Products/PloneMeeting/MeetingConfig.py
@@ -171,6 +171,8 @@ ITEM_WF_STATE_ATTRS = [
     'itemAdviceInvalidateStates',
     'itemAutoSentToOtherMCStates',
     'itemBudgetInfosStates',
+    'itemCommitteesStates',
+    'itemCommitteesViewStates',
     'itemGroupsInChargeStates',
     'itemCopyGroupsStates',
     'itemManualSentToOtherMCStates',
@@ -2502,6 +2504,40 @@ schema = Schema((
         enforceVocabulary=True,
         write_permission="PloneMeeting: Write risky config",
     ),
+    LinesField(
+        name='itemCommitteesStates',
+        widget=MultiSelectionWidget(
+            description="ItemCommitteesStates",
+            description_msgid="item_committees_states_descr",
+            format="checkbox",
+            label='Itemcommitteesstates',
+            label_msgid='PloneMeeting_label_itemCommitteesStates',
+            i18n_domain='PloneMeeting',
+        ),
+        schemata="committees",
+        multiValued=1,
+        vocabulary='listItemStates',
+        default=defValues.itemCommitteesStates,
+        enforceVocabulary=True,
+        write_permission="PloneMeeting: Write risky config",
+    ),
+    LinesField(
+        name='itemCommitteesViewStates',
+        widget=MultiSelectionWidget(
+            description="ItemCommitteesViewStates",
+            description_msgid="item_committees_view_states_descr",
+            format="checkbox",
+            label='Itemcommitteesviewstates',
+            label_msgid='PloneMeeting_label_itemCommitteesViewStates',
+            i18n_domain='PloneMeeting',
+        ),
+        schemata="committees",
+        multiValued=1,
+        vocabulary='listItemStates',
+        default=defValues.itemCommitteesViewStates,
+        enforceVocabulary=True,
+        write_permission="PloneMeeting: Write risky config",
+    ),
     DataGridField(
         name='committees',
         widget=DataGridField._properties['widget'](
@@ -2544,6 +2580,11 @@ schema = Schema((
                                      col_description="committees_supplements_col_description",
                                      vocabulary="listNumbersFromZero",
                                      default='0'),
+                     'enable_editors':
+                        SelectColumn("Committee editors group enabled?",
+                                     col_description="committees_enable_editors_col_description",
+                                     vocabulary="listBooleanVocabulary",
+                                     default='0'),
                      'enabled':
                         SelectColumn("Committee enabled?",
                                      col_description="committees_enabled_col_description",
@@ -2561,7 +2602,7 @@ schema = Schema((
                  'default_assembly', 'default_signatures',
                  'default_attendees', 'default_signatories',
                  'using_groups', 'auto_from',
-                 'supplements', 'enabled'),
+                 'supplements', 'enable_editors', 'enabled'),
         allow_empty_rows=False,
     ),
     BooleanField(
@@ -3457,6 +3498,28 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
             if not v.get('row_id', None):
                 v['row_id'] = 'committee_{0}'.format(self.generateUniqueId())
 
+        # get removed row_ids and remove linked Plone group
+        # rows that were removed
+        storedRowIds = [v['row_id'].strip() for v in self.getCommittees()
+                        if v['enable_editors'] == "1"]
+        rowIds = [v['row_id'].strip() for v in value
+                  if v.get('orderindex_', None) != 'template_row_marker']
+        removedRowIds = [storedRowId for storedRowId in storedRowIds
+                         if storedRowId not in rowIds]
+        # "enable_editors" that was set from "1" to "0"
+        storedEnableEditorsRowIds = [
+            v['row_id'].strip() for v in self.getCommittees()
+            if v['enable_editors'] == "1"]
+        disabledEditorsRowIds = [
+            v['row_id'].strip() for v in value
+            if v.get('orderindex_', None) != 'template_row_marker' or
+            v.get('enable_editors') == "0"]
+        disabledRowIds = [storedRowId for storedRowId in storedEnableEditorsRowIds
+                          if storedRowId not in disabledEditorsRowIds]
+        for row_id_to_remove in tuple(set(removedRowIds + disabledRowIds)):
+            plone_group_id = '{0}_{1}'.format(self.getId(), row_id_to_remove)
+            api.group.delete(plone_group_id)
+
         self.getField('committees').set(self, value, **kwargs)
 
     security.declareProtected(WriteRiskyConfig, 'setMaxShownListings')
@@ -3987,6 +4050,26 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
                 return _('error_committee_row_id_removed_already_used',
                          mapping={'url': brains[0].getURL(),
                                   'committee_label': safe_unicode(self.get_committee_label(removed))})
+            # if "enable_editors", check if linked Plone group is empty
+            if self.getCommittees(committee_id=removed)["enable_editors"] == "1":
+                plone_group_id = '{0}_{1}'.format(self.getId(), removed)
+                if api.group.get(plone_group_id).getGroupMembers():
+                    return translate('committee_removed_plone_group_not_empty',
+                                     domain='PloneMeeting',
+                                     context=self.REQUEST)
+
+        # when changing "enable_editors" from "1" ro "0", Plone group must be empty
+        stored_enable_editors = [row["row_id"] for row in self.getCommittees()
+                                 if row["enable_editors"] == "1"]
+        disabled_editors = [v["row_id"] for v in value
+                            if v.get('orderindex_', None) != 'template_row_marker' and
+                            v["enable_editors"] == "0"]
+        for disabled_editor in set(stored_enable_editors).intersection(disabled_editors):
+            plone_group_id = '{0}_{1}'.format(self.getId(), disabled_editor)
+            if api.group.get(plone_group_id).getGroupMembers():
+                return translate('committee_disabled_editors_plone_group_not_empty',
+                                 domain='PloneMeeting',
+                                 context=self.REQUEST)
 
         # columns using_groups and auto_from are exclusive
         if self.is_committees_using("auto_from", new_value) and \
@@ -6237,6 +6320,29 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
                 self.reindexObjectSecurity()
         return groupIds
 
+    security.declarePrivate('createCommitteeEditorsGroups')
+
+    def createCommitteeEditorsGroups(self, dry_run_return_group_ids=False):
+        '''Create committee Plone groups that will be used to apply
+           the 'MeetingCommitteeEditor' local role on every items of this
+           MeetingConfig regarding self.committees.'''
+        groupIds = []
+        for committee in self.getCommittees():
+            groupSuffix = committee['row_id']
+            # create or update Plone group
+            if committee["enable_editors"] == "1":
+                groupTitleSuffix = translate(
+                    "committee_editors_group_title",
+                    domain="PloneMeeting",
+                    mapping={"label": safe_unicode(committee['label'])},
+                    context=self.REQUEST)
+                groupId, wasCreated = self._createOrUpdatePloneGroup(
+                    groupSuffix,
+                    groupTitleSuffix=groupTitleSuffix,
+                    dry_run_return_group_ids=dry_run_return_group_ids)
+                groupIds.append(groupId)
+        return groupIds
+
     security.declarePrivate('createBudgetImpactEditorsGroup')
 
     def createBudgetImpactEditorsGroup(self, dry_run_return_group_ids=False):
@@ -6316,6 +6422,10 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
         group_ids += self.createPowerObserversGroups(
             force_update_access=force_update_access,
             dry_run_return_group_ids=dry_run_return_group_ids)
+        # Create the corresponding group that will contain committee editors
+        group_ids += self.createCommitteeEditorsGroups(
+            dry_run_return_group_ids=dry_run_return_group_ids)
+        # Custom create groups
         group_ids += self.adapted()._custom_createOrUpdateGroups(
             force_update_access=force_update_access, dry_run_return_group_ids=dry_run_return_group_ids)
         return group_ids
@@ -7062,7 +7172,7 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
 
     security.declarePublic('getCommittees')
 
-    def getCommittees(self, only_enabled=False, **kwargs):
+    def getCommittees(self, only_enabled=False, committee_id=None, **kwargs):
         '''Overrides field 'committees' accessor to be able to pass
            the p_only_enabled parameter that will return
            committees for which enabled is '1'.'''
@@ -7070,6 +7180,12 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
         if only_enabled:
             committees = [committee for committee in committees
                           if committee['enabled'] == '1']
+        # in case we have p_committee_id, only return this element, not a list
+        if committee_id:
+            # manage __suppl__
+            committee_id = committee_id.split('__suppl__')[0]
+            committees = [committee for committee in committees
+                          if committee['row_id'] == committee_id][0]
         return committees
 
     def getCategoriesIds_cachekey(method, self, catType='categories', onlySelectable=True, userId=None):